requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
PySide6>=6.5.0
pystray>=0.19.0
rembg>=2.0.0
